requirements.txt
chromadb>=0.5.15
ollama>=0.3.3
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
pydantic>=2.8.2
python-dotenv>=1.0.1
streamlit>=1.28.0
pypdf2>=3.0.0
python-docx>=0.8.11

